BE + FN Taikhoan + KhachHang

+++ backend/src/main/java/iuh/fit/hotel_booking_backend/service/KhachHangService.java
@@ -14,19 +14,49 @@ public class KhachHangService {
         this.repo = repo;
     }
 
+    //getAll
     public List<KhachHang> getAll() {
         return repo.findAll();
     }
 
+    // getById
     public KhachHang getById(String id) {
         return repo.findById(id).orElse(null);
     }
 
+    // save
     public KhachHang save(KhachHang khachHang) {
+        if (repo.existsById(khachHang.getMaKhachHang())){
+            return null;
+        }
         return repo.save(khachHang);
     }
-
+    // delete
     public void deleteById(String id) {
+        if (!repo.existsById(id)){
+            return ;
+        }
         repo.deleteById(id);
+    }
+
+    // Tìm kiếm theo tên hoặc sdt
+    public List<KhachHang> search(String keyword) {
+        return repo.searchByKeyword(keyword);
+    }
+
+    // update
+    public boolean update(String id, KhachHang updatedKhachHang) {
+        KhachHang existing = repo.findById(id).orElse(null);
+        if (existing == null) return false;
+
+        if (updatedKhachHang.getHoTenKH() != null)
+            existing.setHoTenKH(updatedKhachHang.getHoTenKH());
+        if (updatedKhachHang.getSoDienThoai() != null)
+            existing.setSoDienThoai(updatedKhachHang.getSoDienThoai());
+        if (updatedKhachHang.getDiemTichLuy() > 0)
+            existing.setDiemTichLuy(updatedKhachHang.getDiemTichLuy());
+
+        repo.save(existing);
+        return true;
     }
 }
